requirements.txt
celery==5.0.5
Django==3.2.3
djangorestframework==3.12.4
flake8==3.9.2
redis==3.5.3
psycopg2-binary==2.8.5
PyJWT==2.1.0
Pillow==8.3.1